+++ contracts/samples/TokenPaymaster.sol
@@ -19,7 +19,7 @@ import "../BasePaymaster.sol";
 contract TokenPaymaster is BasePaymaster, ERC20 {
 
     //calculated cost of the postOp
-    uint constant COST_OF_POST = 15000;
+    uint256 constant COST_OF_POST = 15000;
 
     bytes32 immutable public knownWallet;
 
@@ -27,7 +27,7 @@ contract TokenPaymaster is BasePaymaster, ERC20 {
         knownWallet = _knownWallet();
         //make it non-empty
         _mint(address (this),1);
-        approve(owner(), type(uint).max);
+        approve(owner(), type(uint256).max);
     }
 
     // known wallet construct we support the creation of.
@@ -36,20 +36,20 @@ contract TokenPaymaster is BasePaymaster, ERC20 {
     }
 
     //helpers for owner, to mint and withdraw tokens.
-    function mintTokens(address recipient, uint amount) external onlyOwner {
+    function mintTokens(address recipient, uint256 amount) external onlyOwner {
         _mint(recipient, amount);
     }
 
     //TODO: this method assumes a fixed ratio of token-to-eth. subclass should override to supply oracle
     // or a setter.
-    function getTokenToEthOutputPrice(uint valueEth) internal view virtual returns (uint valueToken) {
+    function getTokenToEthOutputPrice(uint256 valueEth) internal view virtual returns (uint256 valueToken) {
         return valueEth / 100;
     }
 
     // verify that the user has enough tokens.
-    function validatePaymasterUserOp(UserOperation calldata userOp, bytes32 /*requestId*/, uint requiredPreFund)
+    function validatePaymasterUserOp(UserOperation calldata userOp, bytes32 /*requestId*/, uint256 requiredPreFund)
     external view override returns (bytes memory context) {
-        uint tokenPrefund = getTokenToEthOutputPrice(requiredPreFund);
+        uint256 tokenPrefund = getTokenToEthOutputPrice(requiredPreFund);
 
         // make sure that verificationGas is high enough to handle postOp
         require(userOp.verificationGas > COST_OF_POST, "TokenPaymaster: gas too low for postOp");
@@ -82,11 +82,11 @@ contract TokenPaymaster is BasePaymaster, ERC20 {
     // this method will be called just after the user's TX with mode==OpSucceeded|OpReverted.
     // BUT: if the user changed its balance in a way that will cause  postOp to revert, then it gets called again, after reverting
     // the user's TX
-    function _postOp(PostOpMode mode, bytes calldata context, uint actualGasCost) internal override {
+    function _postOp(PostOpMode mode, bytes calldata context, uint256 actualGasCost) internal override {
         //we don't really care about the mode, we just pay the gas with the user's tokens.
         (mode);
         address sender = abi.decode(context, (address));
-        uint charge = getTokenToEthOutputPrice(actualGasCost + COST_OF_POST);
+        uint256 charge = getTokenToEthOutputPrice(actualGasCost + COST_OF_POST);
         //actualGasCost is known to be no larger than the above requiredPreFund, so the transfer should succeed.
         _transfer(sender, address(this), charge);
     }
