+++ contracts/test/TestOracle.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.7;
 import "../samples/IOracle.sol";
 
 contract TestOracle is IOracle {
-    function getTokenToEthOutputPrice(uint ethOutput) external pure override returns (uint tokenInput) {
+    function getTokenToEthOutputPrice(uint256 ethOutput) external pure override returns (uint256 tokenInput) {
         return ethOutput * 2;
     }
 }
