+++ contracts/test/TestToken.sol
@@ -8,7 +8,7 @@ contract TestToken is ERC20 {
         ERC20("TST", "TestToken") {
     }
 
-    function mint(address sender, uint amount) external {
+    function mint(address sender, uint256 amount) external {
         _mint(sender, amount);
     }
 }
